requirements.txt
roboflow==1.2.6
python-dotenv==1.1.1
opencv-python-headless==4.10.0.84
Pillow==11.3.0
tqdm==4.67.1
